requirements.txt
google-generativeai
chromadb
sentence-transformers
python-dotenv